week01/requirements.txt
python-dotenv
requests
lxml
pathlib